Separate the request into its own request to the 'Link' queue so that other queues can be added.
