+++ sqs-tester/src/main/java/uk/gov/justice/laa/crime/microservice/sqstester/controller/SqsController.java
@@ -13,7 +13,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import uk.gov.justice.laa.crime.microservice.sqstester.logging.MethodLogger;
-import uk.gov.justice.laa.crime.microservice.sqstester.model.Request;
+import uk.gov.justice.laa.crime.microservice.sqstester.model.requests.LinkRequest;
 import uk.gov.justice.laa.crime.microservice.sqstester.service.MessageQueueProcessor;
 
 @RestController
@@ -31,10 +31,10 @@ public class SqsController {
      * @param request The data we receive in JSON format
      * @return A string with a 200 'OK' status code
      */
-    @PostMapping("/send-message")
+    @PostMapping("/send-message/link")
     @MethodLogger()
-    @Operation(summary = "Send a request off to the message queue.")
-    public ResponseEntity sendMessageToQueue(@Valid @RequestBody Request request) {
+    @Operation(summary = "Send a request off to the 'link' message queue.")
+    public ResponseEntity sendMessageToLinkQueue(@Valid @RequestBody LinkRequest request) {
         try {
             ObjectMapper objectMapper = new ObjectMapper();
             String messageBody = objectMapper.writeValueAsString(request);
